requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
filetype==1.2.0
frozenlist==1.6.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-experimental==0.3.4
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
olefile==0.47
ollama==0.4.8
openai==1.77.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
pdfplumber==0.7.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdf==5.4.0
pytest==8.3.5
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
unstructured==0.17.2
unstructured-client==0.34.0
urllib3==2.4.0
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
fastapi
uvicorn[standard]
pytest-asyncio
